fix codificacion y decodificacion

+++ src/core/huffman/DatoHuffman.cxx
@@ -1,12 +1,12 @@
 #ifndef DATOHUFFMAN_CXX
 #define DATOHUFFMAN_CXX
 
-#include <bits/stdc++.h>
 #include "DatoHuffman.h"
 
-DatoHuffman::DatoHuffman(int c, int n) : c(c), n(n) {}
-DatoHuffman::DatoHuffman()
-{
+#include <bits/stdc++.h>
+
+DatoHuffman::DatoHuffman(int c, unsigned long n) : c(c), n(n) {}
+DatoHuffman::DatoHuffman() {
   c = 0;
   n = 0;
 }
